requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=0.30.1=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
bitsandbytes=0.43.1=pypi_0
blas=1.0=mkl
blinker=1.8.2=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.6.2=hbcca054_0
certifi=2024.6.2=pypi_0
chardet=3.0.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pyhd8ed1ab_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-runtime=12.1.0=0
cudatoolkit=11.7.0=hd8887f6_10
datasets=2.19.1=pypi_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.8=pypi_0
docstring-parser=0.16=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=py310h06a4308_0
flask=3.0.3=pypi_0
flask-cors=4.0.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
googletrans=3.0.0=pypi_0
h11=0.9.0=pypi_0
h2=3.2.0=pypi_0
hpack=3.0.0=pypi_0
hstspreload=2024.6.1=pypi_0
httpcore=0.9.1=pypi_0
httpx=0.13.3=pypi_0
huggingface-hub=0.23.2=pypi_0
hyperframe=5.2.0=pypi_0
idna=2.10=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
imutils=0.5.4=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.3=pyhd33586a_0
ipython=8.25.0=pyh707e725_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py310h06a4308_0
joblib=1.4.2=py310h06a4308_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py310hff52083_0
ld_impl_linux-64=2.38=h1181459_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libffi=3.4.4=h6a678d5_1
libgcc-ng=13.2.0=h77fa898_7
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=13.2.0=h77fa898_7
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py310h5eee18b_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=py310h06a4308_0
nltk=3.8.1=pypi_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
opencv-python=4.10.0.82=pypi_0
openssl=3.3.0=h4ab18f5_3
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
peft=0.11.1=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pip=24.0=py310h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=3.20.3=pypi_0
psutil=5.9.8=py310h2372a71_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=16.1.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pybind11-abi=4=hd3eb1b0_1
pygments=2.18.0=pyhd8ed1ab_0
python=3.10.14=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python_abi=3.10=2_cp310
pytorch=2.3.0=py3.10_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pyttsx3=2.90=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=py310h5eee18b_0
pyzmq=25.1.2=py310h6a678d5_0
readline=8.2=h5eee18b_0
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
rfc3986=1.5.0=pypi_0
rich=13.7.1=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.4.2=py310h1128e8f_1
scipy=1.13.1=py310h5f9d8c6_0
sentencepiece=0.2.0=pypi_0
setuptools=69.5.1=py310h06a4308_0
shtab=1.7.1=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pypi_0
speechrecognition=3.10.4=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
torchtriton=2.3.0=py310
tornado=6.4=py310h2372a71_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.41.2=pypi_0
trl=0.8.6=pypi_0
typing_extensions=4.11.0=py310h06a4308_0
tyro=0.8.4=pypi_0
tzdata=2024.1=pypi_0
unsloth=2024.5=pypi_0
urllib3=2.2.1=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py310h06a4308_0
xformers=0.0.26.post1=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
